% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioseqconvertor_Source.R
\name{DNA}
\alias{DNA}
\title{Define S3 class for DNA sequences}
\usage{
DNA(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA sequence.}
}
\description{
Define S3 class for DNA sequences
}
