% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioseqconvertor_Source.R, R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Convert DNA to RNA}
\usage{
dna_to_rna(dna_sequence)

dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A character string representing a DNA sequence.}
}
\value{
A character string representing the corresponding RNA sequence.

A character string representing the corresponding RNA sequence.
}
\description{
Convert DNA to RNA

Convert DNA sequence to RNA sequence
}
