% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioseqconvertor_Source.R, R/protein_to_cdna.R
\name{protein_to_cdna}
\alias{protein_to_cdna}
\title{Convert Protein to cDNA}
\usage{
protein_to_cdna(protein_seq)

protein_to_cdna(protein_seq)
}
\arguments{
\item{protein_seq}{A character string representing a protein sequence.}
}
\value{
A character string representing the corresponding cDNA sequence.

A character string representing the corresponding cDNA sequence.
}
\description{
Convert Protein to cDNA

Convert protein sequence to cDNA sequence
}
