% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioseqconvertor_Source.R, R/rna_to_protein.R
\name{rna_to_protein}
\alias{rna_to_protein}
\title{Convert RNA to Protein}
\usage{
rna_to_protein(rna_sequence)

rna_to_protein(rna_sequence)
}
\arguments{
\item{rna_sequence}{A character string representing an RNA sequence.}
}
\value{
A character string representing the corresponding protein sequence.

A character string representing the corresponding protein sequence.
}
\description{
Convert RNA to Protein

Convert RNA sequence to protein sequence
}
